Remove superfluous intermediate error state handling
Previously in the bean codebase...

This function was taking a flag to suppress error escalation on the
initial evaluation, likely to account for a previous bug that I'm not
bothering to dig up right now.

If that last sentence sounds like it is full of foreshadowing, you're
right. It is.